requirements.txt
requests==2.32.3
speedtest-cli==2.1.3 